refactor: Create a separate model for StartUp
